backend/backend/requirements.txt
backports.zoneinfo==0.2.1
Django==3.2.18
django-filter==23.1
djangorestframework==3.14.0
drf-base64==2.0
djoser==2.1.0
flake8==5.0.4
gunicorn==20.1.0
importlib-metadata==1.7.0
Markdown==3.3.4
Pillow==9.5.0
psycopg2-binary==2.8.6
python-dotenv==0.20.0
reportlab==3.6.13
requests==2.28.2
requests-oauthlib==1.3.1